Make Page to use AsyncContainer
