+++ src/components/Page.js
@@ -1,17 +1,31 @@
 import React from 'react';
+import { withStyles, useTheme } from '@material-ui/core/styles';
 
 import { useAppContext } from 'context';
+import { AsyncContainer } from 'components';
 
-export default function Page({
-  component: Component = 'div',
-  title = 'Cricket',
-  ...props
-}) {
+export default withStyles((theme) => ({
+  root: {
+    flex: 1,
+    paddingBottom: theme.spacing(10),
+  },
+  contentContainer: {},
+  progressContainer: {
+    position: 'fixed',
+  },
+}))(function Page({ classes, title = 'Cricket', ...props }) {
+  const theme = useTheme();
   const [, updateContext] = useAppContext();
 
   React.useEffect(() => {
     updateContext({ title });
   }, [title, updateContext]);
 
-  return <Component {...props} />;
-}
+  return (
+    <AsyncContainer
+      classes={classes}
+      progressProps={{ size: theme.spacing(10) }}
+      {...props}
+    />
+  );
+});
